Add a version dependency on `w3m`
This is the version I currently have installed, so I'm going to use that
as the dependency. The Melpa version is apparently "20180116.135", but I
guess I'll go with this (taken from my `~/.w3m/.sessions` file).

Presumably my package also works with earlier versions, but this is just
easier than making sure myself.